+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       - HMIPW-DRD3 (Homematic IP Wired Dimming Actuator – 3x channels)
       - HMIPW-DRS8 (Homematic IP Wired Switch Actuator – 8x channels)
       - HMIPW-DRI32 (Homematic IP Wired Inbound module – 32x channels)
+      - HMIP-STV (Inclination and vibration Sensor)
       - Fields
         - connectionType
         - new OptionalFeatures